// src/openapi/model/api-project-to-search.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiProjectToSearch
 */
export interface ApiProjectToSearch {
    /**
     * 
     * @type {string}
     * @memberof ApiProjectToSearch
     */
    keyword?: string;
    /**
     * 
     * @type {number}
     * @memberof ApiProjectToSearch
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiProjectToSearch
     */
    ordererId?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiProjectToSearch
     */
    areaId?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof ApiProjectToSearch
     */
    genreIds?: Array<number>;
    /**
     * 
     * @type {number}
     * @memberof ApiProjectToSearch
     */
    stockFrom?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiProjectToSearch
     */
    stockTo?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiProjectToSearch
     */
    priceFrom?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiProjectToSearch
     */
    priceTo?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiProjectToSearch
     */
    status?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiProjectToSearch
     */
    workDateFrom?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProjectToSearch
     */
    workDateTo?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProjectToSearch
     */
    restraintDateFrom?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProjectToSearch
     */
    restraintDateTo?: string;
}